+++ docs/source/change_log.rst
@@ -12,6 +12,11 @@ Version 4.5.0
 
 * :py:class:`~rule_engine.errors.StringSyntaxError` is now raised for invalid string literals
 * :py:class:`~rule_engine.errors.FunctionCallError` is now raised when a typed function returns an incompatible value
+* Added the new :py:class:`~rule_engine.types.DataType.BYTES` data type
+* Added some new data attributes
+
+    * Added ``starts_with`` and ``ends_with`` to :py:class:`~rule_engine.types.DataType.ARRAY`, and
+      :py:class:`~rule_engine.types.DataType.STRING`
 
 Version 4.4.0
 ^^^^^^^^^^^^^